HH-188149 extract DataSourcePropertiesStorage to separate class
