+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceBeanPostProcessor.java
@@ -11,6 +11,7 @@ import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.core.annotation.AnnotationConfigurationException;
+import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import ru.hh.nab.datasource.DataSourceType;
 
 public class ExecuteOnDataSourceBeanPostProcessor implements BeanPostProcessor {
@@ -40,7 +41,7 @@ public class ExecuteOnDataSourceBeanPostProcessor implements BeanPostProcessor {
 
     Set<String> unconfiguredDataSources = this.affectedBeans.values().stream()
         .flatMap(Collection::stream)
-        .filter(item -> !DataSourceType.isConfigured(item))
+        .filter(item -> !DataSourcePropertiesStorage.isConfigured(item))
         .collect(Collectors.toSet());
     if (dataSourcesAreReady && !unconfiguredDataSources.isEmpty()) {
       String beansString = affectedBeans.entrySet().stream()
